requirements.txt
streamlit
pandas
ortools
pyyaml
fastapi
uvicorn